Switch part of AB calculation back to Time math
The original switch away from Time was to potentially support implicit
coefficients, but I plan to do that in a different way.

It is not clear whether this reduces the floating-point error.  The
conversion from TimeDelta to double has comparable error to the
previous subtractions, but the errors may be correlated in a way that
reduces the overall contribution.  The change is mostly for code
cleanliness.