Make (Gh)ValenciaDivClean Initial Data runtime
Modify the equation of state tag used in evolution to always be a 3d equation of state.
The equation of state must be specified in the input file, but can optionally be
retrieved from the initial data, as opposed to before where the EoS was automatically
retreived from the initial data.